rectoring: create components folders, renamed components files, change exports
