+++ 05 - todo-react-hooks/src/components/Header/index.js
@@ -1,11 +1,9 @@
 import React from "react";
 
-const Header = () => {
-    return(
+export const Header = () => {
+    return (
         <header>
             <h1>To Do List</h1>
         </header>
     );
 }
-
-export default Header;
